requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.1
pillow==10.4.0
pydantic==2.8.0
pydantic_core==2.20.0
replicate==0.27.0
sniffio==1.3.1
typing_extensions==4.12.2
